Created Separate File for VoiceBankHome
