+++ mobile_app/lib/routes.dart
@@ -1,7 +1,7 @@
 import 'package:flutter/material.dart';
 import 'package:voice_banking_poc/screens/landing_screen.dart';
-import 'package:voice_banking_poc/main.dart';
 import 'package:voice_banking_poc/screens/otp_screen.dart';
+import 'package:voice_banking_poc/screens/voice_bank_home.dart';
 
 class MyRouter {
   static Route<dynamic> generateRoute(RouteSettings settings) {
